requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.2
django-cors-headers==3.7.0
django-countries==7.1
djangorestframework==3.12.4
idna==2.10
isbnlib==3.10.7
numpy==1.20.2
opencv-python==4.5.1.48
playsound==1.2.2
pytz==2021.1
pyzbar==0.1.8
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
